fix: address code review iteration 4 findings
Normalize avgjudge i32::MAX sentinel to i64::MAX on DB read to prevent
legacy scores from being misinterpreted as having valid timing data.
Add debug_assert for non-empty gauges in GrooveGauge::new(). Use i64
intermediate in FloatPCM::change_channels to prevent overflow. Add
documentation comments for gauge death-check, SQL ATTACH escaping,
rival clear subtraction safety, and HTML meta-tag parsing fragility.

+++ crates/rubato-audio/src/float_pcm.rs
@@ -208,8 +208,8 @@ impl FloatPCM {
         FloatPCM::new(
             channels,
             self.sample_rate,
-            self.start * channels / self.channels,
-            self.len * channels / self.channels,
+            (self.start as i64 * channels as i64 / self.channels as i64) as i32,
+            (self.len as i64 * channels as i64 / self.channels as i64) as i32,
             samples,
         )
     }
